Introduce variable thresholds for BIP9

+++ src/Stratis.Bitcoin/Base/Deployments/Models/ThresholdStateModel.cs
@@ -34,7 +34,7 @@ namespace Stratis.Bitcoin.Base.Deployments.Models
         public int Height { get; set; }
 
         /// <summary>
-        /// The number of blocks in the each confirmation window.
+        /// The number of blocks in each confirmation window.
         /// </summary>
         [JsonProperty(PropertyName = "confirmationPeriod")]
         public int ConfirmationPeriod { get; set; }
@@ -73,7 +73,7 @@ namespace Stratis.Bitcoin.Base.Deployments.Models
         /// Activation vote threshold for this BIP9 deployment.
         /// </summary>
         [JsonProperty(PropertyName = "threshold")]
-        public int Threshold { get; set; }
+        public long Threshold { get; set; }
 
         /// <summary>
         /// Start time for vote counting for this BIP9 deployment.
